Add files via upload
This repository contains exercises focused on the basics of Python syntax and working with variables. It includes simple programs that demonstrate how to print output, assign different types of data to variables, perform arithmetic operations like addition and multiplication, and calculate averages. It also covers how to check the data type of a variable, print multiple values in a single statement, and handle strings with single or double quotes and escape sequences like backslashes or newlines. Additional concepts include using comments in Python code, assigning and printing None, deleting variables, and checking the memory address of a variable using the id() function. These exercises are designed to help build a strong foundation for anyone starting to learn Python programmin